Moves reduction dispatching into individual reduction files
